+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Apply Bootstrap .table and .table-bordered to all tables
+- Populate CHILD_THEME_NAME, CHILD_THEME_URL, and CHILD_THEME_VERSION based on values in style.css
 
 ## [1.0.0] - 2019-01-31
 ### Added